renamed data-testing/posts.js to data-testing/aws.js
